refactor: split value.rs into modules for better maintainability
Split the 1159-line src/value.rs into a src/value/ module directory
with four files, all under the 500-line convention limit:
- mod.rs (397 lines): Core types (Value, JunctionKind, etc.), PartialEq, utility methods
- display.rs (382 lines): String conversion (to_string_value, version parsing, formatting)
- types.rs (189 lines): Type checking and coercion (eqv, truthy, isa_check, does_check)
- error.rs (205 lines): RuntimeError, RuntimeErrorCode, control flow signals

All public symbols are re-exported so existing imports work unchanged.
